Adding mask to device channel

+++ trunk/devices/src/test/java/net/audumla/devices/io/channel/i2c/rpi/I2ctest.java
@@ -41,7 +41,7 @@ public class I2ctest {
         assert val == d.read();
         d.write((byte) ~val);
         logger.debug("Value : "+d.read());
-        assert ~val == d.read();
+        assert ((byte)~val) == d.read();
     }
 
 }
